requirements.txt
absl-py==0.13.0
backcall==0.2.0
bleach==1.5.0
cached-property==1.5.2
cffi==1.14.6
colorama==0.4.4
cycler==0.10.0
dataclasses==0.8
decorator==5.1.0
h5py==2.10.0
html5lib==0.9999999
importlib-metadata==4.8.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyterplot==0.0.3
Keras==2.1.6
kiwisolver==1.3.1
lrcurve==1.1.0
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.3.56
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
protobuf==3.18.0
PyAudio==0.2.11
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
scipy==1.4.1
six==1.16.0
SoundFile==0.10.3.post1
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
torch==1.8.1
torchaudio==0.8.1
traitlets==4.3.3
typing-extensions==3.10.0.2
wcwidth==0.2.5
Werkzeug==2.0.1
wget==3.2
zipp==3.5.0
